Changed keycloak controller access token logic
